Move marker highlight to ResultItem component
